test: pin PE resource-tree depth/cycle boundaries; fix DAG drop
Address review on the PE-hardening change:
- Make the visited set path-scoped (insert before recursing, remove after)
  so legitimately shared subtables — a DAG, which the format permits — parse
  on every branch instead of the second reference being silently dropped.
  The depth cap remains the hard stack-overflow bound.
- Fix the depth-cap off-by-one: the guard now caps at exactly the three
  conventional levels (type → name → language) instead of allowing a fourth.
- Add the missing positive coverage the guards landed on with none: a legal
  three-level tree parses to its data entry, a fourth level truncates, and a
  shared subtable parses on both branches.
- Correct the comments: three levels is Windows convention, not a spec limit;
  the visited set is path-scoped; the widen-before-add rationale.
- Log truncation and unresolvable-RVA skips (previously silent) via Logger.
- Split the hardening tests into their own file so neither test file trips
  the 400-line --strict file_length cap (the original change pushed
  PEResourceTests to 463, over the limit).
- CHANGELOG: name the three specific crash vectors instead of a blanket claim.
